+++ src/Matchers/PositiveLookBehindMatcher.cpp
@@ -18,7 +18,7 @@ Result *PositiveLookBehindMatcher::match(
         for (int i = start - 1; i >= 0; --i) {
             Result *r = this->matcher->match(matchables, i, previousResults);
             while (r != NULL) {
-                if (r->getLastMatchedIndex() == start - 1) {
+                if (r->getLastMatchedIndex() == (int)(start - 1)) {
                     return new Result(start -1, start - 1);
                 }
                 r = this->matcher->next();
